+++ components/datetime/examples/work_log.rs
@@ -11,7 +11,7 @@ use icu_benchmark_macros::println;
 
 use icu_calendar::{DateTime, Gregorian};
 use icu_datetime::{
-    TypedDateTimeFormatter, neo_marker::NeoYearMonthDayHourMinuteMarker,
+    FixedCalendarDateTimeFormatter, neo_marker::NeoYearMonthDayHourMinuteMarker,
     neo_skeleton::NeoSkeletonLength,
 };
 use icu_locale_core::locale;
@@ -30,11 +30,11 @@ const DATES_ISO: &[(i32, u8, u8, u8, u8, u8)] = &[
 ];
 
 fn main() {
-    let dtf = TypedDateTimeFormatter::<Gregorian, _>::try_new(
+    let dtf = FixedCalendarDateTimeFormatter::<Gregorian, _>::try_new(
         &locale!("en").into(),
         NeoYearMonthDayHourMinuteMarker::with_length(NeoSkeletonLength::Medium),
     )
-    .expect("Failed to create TypedDateTimeFormatter instance.");
+    .expect("Failed to create FixedCalendarDateTimeFormatter instance.");
 
     println!("\n====== Work Log (en) example ============");